test(integer-i64): narrow two i32 round-trip writes in slots.rs
`sequential_file_blocks_read_conflict` writes `f += 0x11223344`
without a cast, then reads 4 bytes and expects round-trip.
Post-round-8.1 (commit `af5fc4e`), plain integer writes emit
8 bytes; on BigEndian the 4-byte read returns all-zero high
bytes, failing the `i32-be-rt` assertion.

Same fix pattern as commit `89d42de` for 20/21-binary.loft:
add `as i32` to both the LittleEndian (line 99) and BigEndian
(line 105) narrow-int writes.  Pure test-source change; no
runtime code touched.

Suite delta: 23 → 22 failing.

+++ tests/slots.rs
@@ -96,13 +96,13 @@ fn sequential_file_blocks_read_conflict() {
   {f = file(\"slots_test_a.bin\"); f#format = LittleEndian;
    assert(f#read(2) as u16 == 0x0302, \"u16-le\"); }
   delete(\"slots_test_a.bin\");
-  {f = file(\"slots_test_a.bin\"); f#format = LittleEndian; f += 0x11223344; }
+  {f = file(\"slots_test_a.bin\"); f#format = LittleEndian; f += 0x11223344 as i32; }
   {f = file(\"slots_test_a.bin\"); f#format = LittleEndian;
    assert(f#read(4) as i32 == 0x11223344, \"i32-le-rt\"); }
   {f = file(\"slots_test_a.bin\"); f#format = BigEndian;
    assert(f#read(4) as i32 == 0x44332211, \"i32-le-as-be\"); }
   delete(\"slots_test_a.bin\");
-  {f = file(\"slots_test_a.bin\"); f#format = BigEndian; f += 0x11223344; }
+  {f = file(\"slots_test_a.bin\"); f#format = BigEndian; f += 0x11223344 as i32; }
   {f = file(\"slots_test_a.bin\"); f#format = BigEndian;
    assert(f#read(4) as i32 == 0x11223344, \"i32-be-rt\"); }
   delete(\"slots_test_a.bin\");
